pyproject.toml
[tool.poetry]
name = "company-revenue-notifier"
version = "0.1.0"
description = ""
authors = ["seriaati <seria.ati@gmail.com>"]
license = "GPL v3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
line-py = {git = "https://github.com/seriaati/line.py"}
fake-useragent = "^1.3.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
tortoise-orm = "^0.20.0"
asyncpg = "^0.29.0"
beautifulsoup4 = "^4.12.2"
seria-library = {git = "https://github.com/seriaati/seria-library"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
